pyproject.toml
[project]
name = "glasslab-metric-search"
version = "0.2.0"
description = "Contrastive representation learning for unseen class generalization on Glasslab"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
  # Core ML frameworks
  "PyYAML>=6.0.2",
  "torch>=2.1.0",
  "torchvision>=0.16.0",
  "timm>=1.0.0",
  # JAX for composable transformations
  "jax>=0.4.30",
  "jaxlib>=0.4.30",
  "optax>=0.2.0",
  # Deep Metric Learning (DML)
  "pytorch-metric-learning>=2.6.0",
  "faiss-cpu>=1.8.0",
  "umap-learn>=0.5.0",
  "scikit-learn>=1.3.0",
  "optuna>=3.4.0",
  "mlxtend>=0.22.0",
  "torchcp>=0.2.0",
  # AutoML orchestrators
  "auto-sklearn>=0.14.0",
  "autogluon.core>=1.0.0",
  "flaml>=2.2.0",
  "tpot>=0.12.0",
  "h2o>=3.44.0",
  "pycaret>=3.2.0",
  # Hyperparameter optimization
  "syne-tune>=0.5.0",
  "ray[tune]>=2.35.0",
  # Uncertainty quantification
  "uncertainty-toolbox>=2023.12.3",
  "puncc>=1.0.0",
  "statsforecast>=1.7.0",
  # Conformal prediction alternatives
  "sklearn-conformal>=1.1.0",
]

[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["search*", "src*", "benchmarks*", "scripts*"]

[tool.setuptools.package-data]
src = ["py.typed"]
